Fix sorting of modules when the dependency graph changes between branch builds
Previously we were only sorting based on the dependency graph of the current
build for the first module, but different modules have have different
current branch builds. This change sorts the modules in each branch build
separately.
